Refactor communication into a component held yb an IM/Vehicle
This replaces inheritance of a CommunicativeAgent, with composition of a
MessagingUnit. This should make the framework more extensible, and allow
for the modelling, for example, of communication imperfections by providing
different implementations of the MessagingUnit class.

Provide a basic implementation, DistanceBasedUnit, which allows for the sending
of messages to vehicles and infrastructure within a fixed radius, defined
by a communication_range parameter. Use this MessagingUnit in the QBIM algo
example.

Closes gh-2
'