Fix bug in Intersection.get_tiles_for_vehicle() using Shapely
There was a bug discovered when using a qb_im discretisation of 1,
where tiles would not be identified when the vehicle was entirely
contained in the tile. Because previously, we were checking whether
any of the corners of the tile was inside the vehicle, which doesn't
have to be the case for them to intersect.

Added Shapely dependency to do this for us, using their handy
Polygon.intersects() method.